Add base crawler for merdeka.com. Refactor merdeka.com for jabar and jatim.
